refactor: mv PERMIT_PRIMARY_TYPE -> new constants/signatures.ts
