+++ shared/constants/transaction.ts
@@ -195,19 +195,3 @@ export enum TokenStandard {
   /** Not a token, but rather the base asset of the selected chain. */
   none = 'NONE',
 }
-
-/**
- * EIP-712 Permit PrimaryTypes
- */
-export enum PERMIT_PRIMARY_TYPE {
-  PERMIT = 'Permit',
-  PERMIT_BATCH = 'PermitBatch',
-  PERMIT_BATCH_TRANSFER_FROM = 'PermitBatchTransferFrom',
-  PERMIT_SINGLE = 'PermitSingle',
-  PERMIT_TRANSFER_FROM = 'PermitTransferFrom',
-  ORDER = 'Order',
-  ORDER_COMPONENTS = 'OrderComponents',
-}
-
-export const PERMIT_PRIMARY_TYPES: string[] =
-  Object.values(PERMIT_PRIMARY_TYPE);
