+++ ui/pages/confirmations/utils/confirm.ts
@@ -2,7 +2,7 @@ import { ApprovalRequest } from '@metamask/approval-controller';
 import { ApprovalType } from '@metamask/controller-utils';
 import { TransactionType } from '@metamask/transaction-controller';
 import { Json } from '@metamask/utils';
-import { PERMIT_PRIMARY_TYPES } from '../../../../shared/constants/transaction';
+import { PERMIT_PRIMARY_TYPES } from '../../../../shared/constants/signatures';
 import { parseTypedDataMessage } from '../../../../shared/modules/transaction.utils';
 import { sanitizeMessage } from '../../../helpers/utils/util';
 import { SignatureRequestType } from '../types/confirm';
